% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{ggtern_base}
\alias{ggtern_base}
\title{Create a base ggtern object to plot over}
\usage{
ggtern_base(
  x_color = "#de8a5a",
  y_color = "#CFC994",
  z_color = "#008080",
  x_label = "Low",
  y_label = "Medium",
  z_label = "High"
)
}
\arguments{
\item{x_color}{the color of the 'x' segment. Defaults to "#de8a5a"}

\item{y_color}{the color of the 'y' segment. Defaults to "#CFC994"}

\item{z_color}{the color of the 'z' segment. Defaults to "#008080"}

\item{x_label}{the name of the 'x' label. Defaults to "Low"}

\item{y_label}{the name of the 'y' label. Defaults to "Medium"}

\item{z_label}{the name of the 'z' label. Defaults to "High"}
}
\description{
This function creates a ternary plot base using ggtern. It creates the plot and creates three segments that easilly separate zones in the data, making the chart easier to interpret
}
\examples{
ggtern_base()
}
