% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{jll_cols}
\alias{jll_cols}
\title{Apply the JLL palette to a plot}
\usage{
jll_cols(...)
}
\description{
This function returns the colors in the JLL palettes. It is used by the scale_color_jll() functions, and rarely needs to be called directly.
}
\examples{
jll_cols()
}
