% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{jll_pal}
\alias{jll_pal}
\title{Apply the JLL palette to a plot}
\usage{
jll_pal(palette = "primary", reverse = FALSE, ...)
}
\description{
This function generates the JLL color palettes. It is used by the scale_color_jll() functions, and rarely needs to be called directly.
}
\examples{
jll_pal()
}
