% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{jll_palette_cols}
\alias{jll_palette_cols}
\title{Apply the JLL palette to a plot}
\usage{
jll_palette_cols(palette = "data_viz", n = FALSE)
}
\description{
This function retrieves the colors within one of the JLL palettes. It is useful when applying a subset of the palette to a particular plot.
}
\examples{

}
