% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLLify-package.R
\docType{package}
\name{JLLify-package}
\alias{JLLify}
\alias{JLLify-package}
\title{JLLify: Apply a JLL style to ggplot objects}
\description{
This package contains a number of functions to apply JLL's corporate palettes to a ggplot object. It also provides functions to apply a clean and modern look to graphs, in both a light and dark theme.
}
\author{
\strong{Maintainer}: Owen King \email{owen.king@eu.jll.com}

}
\keyword{internal}
