% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{scale_colour_jll}
\alias{scale_colour_jll}
\title{Apply JLL colors to a plot}
\usage{
scale_colour_jll(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{the name of the color palette you wish to apply, defaults to "primary". Other supported palettes are "primary_3", "primary_3_dark", "secondary", "colorfull", "primary_secondary" and "ten_cols"}

\item{discrete}{whether the variable the palette being applied to is discrete. Defaults to TRUE.}

\item{reverse}{whether to reverse the color palette. Defaults to FALSE.}
}
\description{
This function applies one of JLL's corporate color palletes to a ggplot object. It is the same as scale_color_jll(), but with the English spelling of "colour".
}
\examples{
scale_colour_jll()
}
