% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{theme_jll_modern}
\alias{theme_jll_modern}
\title{Modernise the look of a ggplot object}
\usage{
theme_jll_modern(...)
}
\description{
This function applies a modern style to a ggplot object.
}
\examples{
theme_jll_modern()
}
