% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{theme_jll_modern_dark}
\alias{theme_jll_modern_dark}
\title{Modernise the look of a ggplot object}
\usage{
theme_jll_modern_dark(...)
}
\description{
This function applies a modern dark-theme style to a ggplot object.
}
\examples{
theme_jll_modern_dark()
}
