% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{theme_map}
\alias{theme_map}
\title{Style a map with a minimal look}
\usage{
theme_map(...)
}
\description{
This function applies a clean theme to a map. It removes all grids and axis markings.
}
\examples{
theme_map()
}
