% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{theme_map_dark}
\alias{theme_map_dark}
\title{Style a map with a minimal look}
\usage{
theme_map_dark(...)
}
\description{
This function applies a clean dark-theme to a map. It removes all grids and axis markings.
}
\examples{
theme_map_dark()
}
