% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jll.R
\name{theme_ternary_dark}
\alias{theme_ternary_dark}
\title{Style a the ternary plot}
\usage{
theme_ternary_dark(
  plot,
  x_color = "#de8a5a",
  y_color = "#CFC994",
  z_color = "#008080",
  ...
)
}
\arguments{
\item{x_color}{the color of the 'x' segment. Defaults to "#de8a5a"}

\item{y_color}{the color of the 'y' segment. Defaults to "#CFC994"}

\item{z_color}{the color of the 'z' segment. Defaults to "#008080"}
}
\description{
This function applies a modern look to a ggtern object
}
\examples{
theme_ternary_dark()
}
